+++ app/components/views/Offset/ProjectSearch/index.tsx
@@ -1,18 +1,9 @@
 import React, { FC, useState } from "react";
-import { Text } from "@klimadao/lib/components";
-import { ButtonPrimary } from "@klimadao/lib/components";
 
-import { ProjectSearchFilter } from "./ProjectSearchFilter";
-import { SelectProjectButton } from "./SelectProjectButton";
+import { CarbonProject } from "../SelectiveRetirement/queryProjectDetails";
+import { ProjectSearchForm } from "./ProjectSearchForm";
+import { ProjectSelection } from "./ProjectSelection";
 import { ConfirmProjectSelection } from "./ConfirmProjectSelection";
-
-import {
-  types,
-  countries,
-  vintages,
-  mockProjectDetails,
-} from "./filterOptions";
-import * as styles from "./styles";
 
 type Props = {
   setIsLoading: (boolean: boolean) => void;
@@ -21,92 +12,37 @@ type Props = {
 
 export const ProjectSearch: FC<Props> = (props) => {
   const [step, setStep] = useState("search");
-
-  const [currentFilter, setCurrentFilter] = useState(null);
-  const [type, setType] = useState(null);
-  const [region, setRegion] = useState(null);
-  const [vintage, setVintage] = useState(null);
-  const [projects, setProjects] = useState(null);
-  const [selectedProject, setSelectedProject] = useState(null);
-
-  const handleSubmit = () => {
-    console.log([type, region, vintage]);
-    props.setIsLoading(true);
-
-    setTimeout(() => {
-      setStep("selectProject");
-      setProjects(mockProjectDetails);
-      props.setIsLoading(false);
-    }, 1500);
-  };
+  const [projects, setProjects] = useState<CarbonProject[] | null>(null);
+  const [selectedProject, setSelectedProject] = useState<CarbonProject | null>(
+    null
+  );
 
   return (
     <>
       {step === "search" && (
-        <>
-          <div>
-            <ProjectSearchFilter
-              name="type"
-              options={types}
-              isOpen={currentFilter === "type"}
-              onClick={setCurrentFilter}
-              onChange={setType}
-            />
-            <ProjectSearchFilter
-              name="region"
-              options={countries}
-              isOpen={currentFilter === "region"}
-              onClick={setCurrentFilter}
-              onChange={setRegion}
-            />
-            <ProjectSearchFilter
-              name="vintage"
-              options={vintages}
-              isOpen={currentFilter === "vintage"}
-              onClick={setCurrentFilter}
-              onChange={setVintage}
-            />
-          </div>
-
-          <ButtonPrimary label="Find project" onClick={handleSubmit} />
-        </>
+        <ProjectSearchForm
+          setStep={setStep}
+          setProjects={setProjects}
+          setIsLoading={props.setIsLoading}
+        />
       )}
 
       {step === "selectProject" && projects && (
-        <div className={styles.selectProjectContainer}>
-          <Text t="caption" color="lighter">
-            Select a project to continue
-          </Text>
-          <div className={styles.projectList}>
-            {projects.map((project, index) => (
-              <SelectProjectButton
-                key={index}
-                project={project}
-                setSelectedProject={setSelectedProject}
-                selected={true}
-                active={selectedProject?.tokenAddress === project.tokenAddress}
-              />
-            ))}
-          </div>
-
-          <div className={styles.projectActionButtons}>
-            <ButtonPrimary
-              label="Back to filters"
-              onClick={() => setStep("search")}
-            />
-            <ButtonPrimary
-              label="Select project"
-              onClick={() => setStep("confirmProject")}
-              disabled={!selectedProject}
-            />
-          </div>
-        </div>
+        <ProjectSelection
+          projects={projects}
+          selectedProject={selectedProject}
+          setProjects={setProjects}
+          setSelectedProject={setSelectedProject}
+          setStep={setStep}
+        />
       )}
 
       {step === "confirmProject" && selectedProject && (
         <ConfirmProjectSelection
-          setStep={setStep}
           projectName={selectedProject.name}
+          projectAddress={selectedProject.tokenAddress}
+          setStep={setStep}
+          setProjectAddress={props.setProjectAddress}
         />
       )}
     </>
