--- app/components/views/Offset/ProjectSearch/styles.ts
@@ -1,70 +1,0 @@
-import { css } from "@emotion/css";
-
-export const projectList = css`
-  display: flex;
-  flex-direction: column;
-  gap: 1.2rem;
-  max-height: 38rem;
-  overflow-y: scroll;
-
-  ::-webkit-scrollbar {
-    display: none;
-  }
-`;
-
-export const projectActionButtons = css`
-  display: flex;
-  gap: 0.8rem;
-  justify-content: space-around;
-  width: inherit;
-
-  button {
-    width: 100%;
-
-    :first-of-type {
-      background-color: var(--surface-01);
-    }
-  }
-`;
-
-export const selectProjectContainer = css`
-  position: relative;
-  display: inherit;
-  flex-direction: inherit;
-  gap: inherit;
-`;
-
-export const confirmSelection = css`
-  height: 48rem;
-  padding: 2rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  gap: 2rem;
-  background-color: var(--surface-02);
-  border: 0.2rem solid var(--surface-01);
-  border-radius: 0.8rem;
-
-  button {
-    width: 100%;
-
-    :last-of-type {
-      background-color: var(--surface-01);
-    }
-  }
-`;
-
-export const icon = css`
-  min-height: 4rem;
-  min-width: 4rem;
-  fill: var(--klima-green);
-`;
-
-export const buttons = css`
-  width: 100%;
-  margin-top: 3.2rem;
-  gap: 1.2rem;
-  display: flex;
-  flex-direction: column;
-`;
